Create a MapSetting for APIService

+++ app/src/main/java/com/example/uit_project/model/datapoint/Datapoint.java
@@ -8,13 +8,13 @@ import java.util.Arrays;
 public class Datapoint {
 
     private long x;
-    private double y;
+    private float y;
 
     public long getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 }
